(#1276) rename to choria governor, some bug fixes
